backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
python-dotenv
redis
openai
django-filter
drf-spectacular